requirements_onnx.txt
onnxruntime-gpu==1.14
Pillow
matplotlib
tqdm
scipy
opencv-python-headless